# backend/docs/swagger.yaml
basePath: /v1
definitions:
  auth.AuthResponse:
    properties:
      role:
        type: string
      tenant_id:
        type: string
      token:
        type: string
      user_id:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.SignupRequest:
    properties:
      business_name:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  billing.PlanResponse:
    properties:
      billing_interval:
        example: monthly
        type: string
        x-nullable: true
      max_orders_per_month:
        example: 100
        type: integer
        x-nullable: true
      max_products:
        example: 50
        type: integer
        x-nullable: true
      max_users:
        description: null = unlimited
        example: 3
        type: integer
        x-nullable: true
      name:
        example: free
        type: string
      price_amount:
        example: 0
        type: number
      price_currency:
        example: NGN
        type: string
    type: object
  billing.UsageResponse:
    properties:
      orders_cap:
        example: 100
        type: integer
        x-nullable: true
      orders_used:
        example: 42
        type: integer
      plan_name:
        example: free
        type: string
      products_cap:
        example: 50
        type: integer
        x-nullable: true
      products_used:
        example: 12
        type: integer
      users_cap:
        example: 3
        type: integer
        x-nullable: true
      users_used:
        example: 2
        type: integer
    type: object
  finance.InvoiceResponse:
    properties:
      amount:
        example: 1500.5
        type: number
      invoice_id:
        example: e5f6g7h8-1234-5678-90ab-cdef12345678
        type: string
      issued_at:
        example: "2026-07-12T21:15:35Z"
        type: string
      order_id:
        example: d3b07384-d113-4956-953e-52f01f05e3d9
        type: string
      status:
        description: e.g., pending, paid, cancelled
        example: paid
        type: string
    type: object
  httputil.APIError:
    properties:
      error:
        example: invalid request body
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Multi-tenant REP system for OneDesk.
  title: OneDesk API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Verifies user credentials and issues a JWT access token.
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: invalid request body
          schema:
            $ref: '#/definitions/httputil.APIError'
        "401":
          description: invalid email or password
          schema:
            $ref: '#/definitions/httputil.APIError'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/httputil.APIError'
      summary: Authenticate user
      tags:
      - auth
  /billing/usage:
    get:
      description: Fetches real-time transaction, item tracking, and staff identity
        usage statistics against plan cap thresholds.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/billing.UsageResponse'
        "401":
          description: missing valid tenant validation credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "402":
          description: plan metric consumption limitation ceiling reached
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: internal service persistence layer failure
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get tenant quota consumption
      tags:
      - billing
  /plans:
    get:
      description: Returns a matrix layout of all structural subscription plan capabilities
        and resource limits.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/billing.PlanResponse'
            type: array
        "500":
          description: internal server error context details
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List pricing tier catalogs
      tags:
      - billing
  /tenants:
    post:
      consumes:
      - application/json
      description: Creates a new business tenant profile along with its initial root
        admin account.
      parameters:
      - description: Signup Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: invalid request body / validation error
          schema:
            $ref: '#/definitions/httputil.APIError'
        "409":
          description: email already registered
          schema:
            $ref: '#/definitions/httputil.APIError'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/httputil.APIError'
      summary: Register a new tenant
      tags:
      - auth
...
  /v1/invoices:
    get:
      description: Returns all issued invoices bound to the current tenant context.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/finance.InvoiceResponse'
            type: array
        "401":
          description: Missing valid tenant credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal service database ledger query fault
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List tenant invoice ledger records
      tags:
      - finance
  /v1/invoices/{invoiceId}:
    get:
      description: Retrieves structural billing components, settlement balance quantities,
        and paid statuses for a specific invoice.
      parameters:
      - description: System invoice tracking index token (UUID)
        in: path
        name: invoiceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/finance.InvoiceResponse'
        "400":
          description: Malformed route sequence path ID value format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing valid tenant credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No invoice record matches the specified balance trace token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal service ledger entity retrieval database error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get details of an invoice
      tags:
      - finance
  /v1/invoices/{invoiceId}/pay:
    patch:
      description: Transitions an open invoice state to paid.
      parameters:
      - description: System invoice tracking index token (UUID)
        in: path
        name: invoiceId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/finance.InvoiceResponse'
        "400":
          description: Malformed route sequence path ID value format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Missing valid tenant credentials
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: No invoice record matches the specified balance trace token
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal service balance modification execution pipeline error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Process invoice settlement payment
      tags:
      - finance
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"